プロフィール部分の要素があるか確認するテストを追加した。
